requirements.txt
gevent
scikit-image
opencv-python
tensorboard==2.1.1
tensorflow-estimator==2.1.0
absl-py==0.9.0
astor==0.8.1
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cycler==0.10.0
Flask==1.1.1
gast==0.2.2
google-auth==1.11.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.27.2
h5py==2.10.0
idna==2.8
imutils==0.5.3
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.2.1
MarkupSafe==1.1.1
numpy==1.18.1
oauthlib==3.1.0
opt-einsum==3.1.0
pandas==1.0.1
Pillow==7.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
six==1.14.0
termcolor==1.1.0
tqdm==4.42.1
urllib3==1.25.8
Werkzeug==1.0.0
wrapt==1.11.2
